Complete 64 bit changes for window numbers for X windows
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gui/branches/gnustep_testplant_branch@40206 72102866-910b-0410-8b05-ffd578937521

+++ Headers/Additions/GNUstepGUI/GSDisplayServer.h
@@ -127,7 +127,7 @@ APPKIT_EXPORT NSString *GSScreenNumber;
 - (NSInteger) window: (NSRect)frame : (NSBackingStoreType)type : (unsigned int)style
 	      : (int)screen;
 - (void) termwindow: (NSInteger)win;
-- (int) nativeWindow: (void *)winref : (NSRect*)frame : (NSBackingStoreType*)type 
+- (NSInteger) nativeWindow: (void *)winref : (NSRect*)frame : (NSBackingStoreType*)type
                     : (unsigned int*)style : (int*)screen;
 
 /* Only if handlesWindowDecorations returns YES. */
