+++ Source/GSDisplayServer.m
@@ -572,7 +572,7 @@ GSCurrentServer(void)
 
 /** Create all the backend structures for a reference to a native window and 
     return the extend, backing type, style and screen for that window. */ 
-- (int) nativeWindow: (void *)winref
+- (NSInteger) nativeWindow: (void *)winref
 		    : (NSRect*)frame
 		    : (NSBackingStoreType*)type 
 		    : (unsigned int*)style
